Post Project Update
I neglected to upload Revision B at the time I generated it for the Requirements, as I was focused on completing the project itself. My original plan was to upload revisions, and my first version of the report after I completed them for testing, but customer requested I take on a second project, which they wanted to bypass documentation for, which resulted in the "Third Party Submitted by User Without Rejections and Reversals v2.sql" file instead of uploading and managing documentation.